+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to
 - ✨(frontend) introduce an "unauthenticated" participant badge
 - ✨(backend) add roomkit viewset to start a room without WebRTC join
 - ✨(frontend) let users set default configuration for generated links
+- ✨(frontend) expose media state to external gateways
 
 ### Changed